+++ Examples/WarpTensorImageMultiTransform.cxx
@@ -542,7 +542,7 @@ void WarpImageMultiTransform(char *moving_image_filename, char *output_image_fil
                              TRAN_OPT_QUEUE & opt_queue, MISC_OPT & misc_opt)
 {
   // typedef itk::Vector<float,6> PixelType;
-  typedef itk::SymmetricSecondRankTensor<double, 3> PixelType;
+  typedef itk::SymmetricSecondRankTensor<float, 3> PixelType;
   typedef itk::Image<PixelType,
                      ImageDimension>                                                               TensorImageType;
   typedef itk::Image<float, ImageDimension> ImageType;
